Add: suggested changes and additional improvements
devel doesn't always have the most pertinent development code. Replaced with general references so that the contributor can decide which branch they are interested in. 

Didn't feel clean to invade the getting started list. It was just so tidy. There is now a development header that lists useful information for starting project development. This may or may not be what we want.